+++ lib/src/solid/constants/common.dart
@@ -116,6 +116,7 @@ const String termsPrefix = 'terms:';
 const String filePrefix = 'file:';
 const String dataPrefix = 'data:';
 const String resIdPrefix = 'resourceId:';
+const String logIdPrefix = 'logId:';
 
 /// String variables for creating files and directories on solid server
